+++ src/Recyclarr.Cli/Console/Helpers/ProviderProgressHandler.cs
@@ -1,4 +1,3 @@
-using Recyclarr.Cli.Logging;
 using Recyclarr.ResourceProviders.Infrastructure;
 using Recyclarr.ResourceProviders.Storage;
 using Spectre.Console;
@@ -7,33 +6,23 @@ namespace Recyclarr.Cli.Console.Helpers;
 
 internal class ProviderProgressHandler(IAnsiConsole console, ProviderInitializationFactory factory)
 {
-    public async Task InitializeProvidersAsync(
-        ConsoleOutputSettings outputSettings,
-        CancellationToken ct
-    )
+    public async Task InitializeProvidersAsync(CancellationToken ct)
     {
-        if (!outputSettings.IsRawOutputEnabled)
-        {
-            await console
-                .Progress()
-                .AutoClear(false)
-                .Columns(
-                    new TaskDescriptionColumn(),
-                    new ProgressBarColumn(),
-                    new PercentageColumn(),
-                    new SpinnerColumn()
-                )
-                .StartAsync(async ctx =>
-                {
-                    var task = ctx.AddTask("Initializing Resource Providers");
-                    var progress = new ProgressReporter(console, task);
-                    await factory.InitializeProvidersAsync(progress, ct);
-                });
-        }
-        else
-        {
-            await factory.InitializeProvidersAsync(null, ct);
-        }
+        await console
+            .Progress()
+            .AutoClear(false)
+            .Columns(
+                new TaskDescriptionColumn(),
+                new ProgressBarColumn(),
+                new PercentageColumn(),
+                new SpinnerColumn()
+            )
+            .StartAsync(async ctx =>
+            {
+                var task = ctx.AddTask("Initializing Resource Providers");
+                var progress = new ProgressReporter(console, task);
+                await factory.InitializeProvidersAsync(progress, ct);
+            });
     }
 
     private class ProgressReporter(IAnsiConsole console, ProgressTask task)