+++ src/Recyclarr.Cli/Logging/IConsoleOutputSettings.cs
@@ -1,6 +1,0 @@
-namespace Recyclarr.Cli.Logging;
-
-internal interface IConsoleOutputSettings
-{
-    bool IsRawOutputEnabled { get; }
-}